Add catalogue entries
This is the initial conversion from the Transmittal files. Only minimal
work has been done, enough to make it a valid yaml file. Further work
will need to be done to break out dimensions and numeric dates, to
properly name the fields, etc. There are also a number of oddities to
be aware of, which we’ll need to address:

- Four of the cat numbers are not straight integers (146 bis, 252 bis,
626a, and 626b)
- There is no cat 456
- There is a TBD cat entry at the end
- Cat 622 is currently a double entry (with two descriptions, two
dimensions, etc.) that will need to be combined.
- Cat 52 and 352 have a footnote in their Description fields
- Field names are consistent except “Condition and Fabric” and
“Condition” are used variously.
- There are a few multi-paragraph Descriptions, Discussions, and
Parallels. They can be found in the document by searching for a pipe “|”
- Provenance and Bibliography info will need to be added/updated from
TMS, and a Place of Manufacture or Origin ID will be added.