Add update files
Includes additional numbering for the Bussiere lamps in Loeschcke Type
VIII on the typological classification. And in the Intros and Entries,
I needed to replace all the lowercase letters that were formats as All
Caps, with actual capitals with no formatting. Pandoc conversion to
markdown left them all as lowercase.