Build out entry layout
This commit is another milestone in getting the entry layout working
properly. It adds:

* Images for Cats 1-30 (some are missing) for main view (sized for use
  as a cover image), and thumbnail view (top view, small size for grid
  images).
* Details component: slides in from edges of screen when triggered by
  clicking on a grid item in the catalogue entries grid of the entry
  layout. In the future this will accommodate a leaflet zoom instance.

In YAML front matter for catalogue entries, a cover number should be
specified: this will grab an image of the same catalogue number from
assets/images/main.

Finally, the cat_scale and cat_image partials have been deleted since
they are not going to be used for this particular book.