Implement basic deepzoom + JS cleanup
This commit strips out a lot of the un-used javascript for now, since
there was a lot carried over from the previous project. Some may get
added back in as needed.

The leaflet-deepzoom plugin is being used to handle deep-zoom image
handling (this also makes the library compatible with images sliced in
the way that the DZT tool operates).

Currently this code all lives in the UI functions file. This code
probably needs to be reorganized, and potentially the entire JS
packaging system for this app could be updated (module loading would
greatly improve things, for ex).