Basic Gulp asset pipeline (SCSS only)
This commit replaces the sprockets asset pipeline (which is becoming
unreliable) with a Javascript pipeline based on Gulp. Right now this
only handles stylesheets, but in the current state the script can
produce the same output as under Ruby (using Bourbon, Neat, etc).

The Bourbon 4.2 and Neat 1.8 stylesheets have been hard-coded into the
repo and added to source since gulp-sass won't be relying on the gem
version of these libraries. Some adjustments in stylesheets needed to be
made to no longer rely on ruby features (neat-helpers and font import
code, primarily).

A basic eslintrc.js file has been added to the project to streamline
future JS development (code will need to be switched to modular format
going forward, but this will be a welcome change).