Improved webpack & JS configuration
This tweaks the ESLINT config to align more closely with the projects
style conventions (finally can say that JS tooling is working with and
not against me).

One important note about both eslint config as well as webpack is the
allowance made for jQuery as a global variable. This project stands at a
transition of two different approaches of writing javascript code: it's
getting complex enough that the old-school concatenation of files
approach is hitting its limits, but legacy jQuery plugins are still
relied upon quite heavily.

As a compromise, the modular style of explicit require() statements will
be used in application code wherever possible, but some of the external
libraries (jquery and jq plugins mainly) will remain available as
globals. Because handling all of this correctly in webpack is pretty
tricky, I'm just goign to bypass webpack for all global libraries for
now: they will be pulled in via separate script tags in the top-level
page template. A few extra HTTP requests seems like a small price to pay
for sanity at this point.