Finalize asset pipeline (basic) in Gulp
This finishes the basic migration of the asset pipeline over to Gulp
instead of Sprockets. This only handles JS and SCSS files (html
templates, markdown rendering, etc. is still done in Ruby), but it is
sufficient for the purposes of this project. Webpack is used to allow
the use of require() statements and modular code style, but otherwise no
exotic transiplation is going on.

Going forward, the contents of the javascripts/vendor folder should be
gradually removed so that NPM can manage these dependencies. This could
potentially be done with Bourbon as well but for now leaving that alone.

Leaflet works well in the CommonJS format (var L = require("leaflet")),
but the leaflet-deepzoom plugin had to be reworked slightly. Fortunately
since this is only a single JS file without too much code, re-working
was pretty easy. Going to treat this file like source code instead of an
external dependency going forward since it has deviated from the
original version on GitHub.