Re-enable catalogue extensions
More refactoring and clean-up is in order here, but this commit
re-enables the catalogue extension and associated helper methods from
Roman Mosaics. The manifest file generated from the image resizer
program has been added as plates.yml, which is used to generate a JSON
file which can be queried on the front-end to populate leaflet
instances. The catalogue extension has been reorganized somewhat for
clarity. Also, for now the ZipFileGenerator has been disabled, not sure
if this is going to be needed going forward.