Catalogue page design improvements
This commit modifies the catalogue page so that instead of a 50/50
vertical split on large screens, more space (65%) is allocated for the
zoomable image component, with the remaining used for the tombstone
text.

In addition, this commit addresses one of the long-standing visual bugs
for this type of design, which is the dreaded double-scrollbar. To make
the modal experience more seamless, a .noscroll class is added to the
body of the page whenever the modal details view is open. This disables
scroll on the layer underneath the modal, so that the viewer only sees a
scrollbar for the content they are currently interacting with.