Build out UI class
AJAX and data processing with lodash is now working. One thing that
seems to matter now, which was not an obvious issue in previous
versions: lodash will strictly distinguish between: 2 and "2" when
searching through a collection: a CatNumCheck() method has been added to
the class that coerces stringified numbers into proper integers, but
leaves non-numerical values as they were.

In most places jQuery dependencies have been removed, but $.get() is way
too useful/convenient to strip out for now. Will see about switching out
for XMLHttpRequest or similar later on.