Improved webpack conf
The ignore line in config rb is now properly ignoring all files except
for the application.js output (which is also now minified).

Jquery has been added as a dependency and exposed to modules as a global
variable. This means that webpack can now handle importing SmoothState,
Velocity, etc. as well and there is no need for extra script tags in the
templates.