Basic Search component
This adds a basic Search class and some UI wiring to the app. Lunr JS is
being used for client-side search, and the search.rb extension has been
re-enabled to generate a custom index out of the catalogue data
file (only some fields are getting picked up here).

The next step is to combine the search feature with the lookup data file
so that each catalogue entry can be linked to whatever page it occurrs on.