Basic EPUB export (mostly) working
This commit reorganizes the extensions code significantly, replacing the
old catalogue.rb extensions with the newer book.rb extension and
friends.

A lot of custom modifications have been added to this for now so that
things work with the complicated structure of this particular book.

To generate an EPUB, run `middleman build -e epub` like normal. For some
reason EPZIP is not working programatically, so you need to manuall run
`epzip dist/epub dist/book.epub` to zip up the files. Images need to be
better handled as well. Will return to this once the basics of the web
version have been pinned down better.