Remove {{loc:}} indicators and replace with links
This was a global find-replace powered by this regular expression:
{{(loc_[\d]+):([\s\S]+?)}}

The change replaces all of the "location tags" in the source
files ("{{figure}}" and "{{insc}}" still exist in a few places) and
replaces them with relative links. These links assume the URL structure
of the web-based version of the book and will probably need to be
stripped out in other versions (like Epub, which fails to validate if
there are broken links).

Next step is to implement a hash-listening function for the map instance
to respond when the user attempts to navigate to these links.