Reduce redundant search.json ajax requests
This is a workaround to reduce the number of requests for search data
made by the UI class on initialization. Previously this was happening
every time a new page was visited via smoothstate. Ideally it should
only happen on the inital load since the data is static.

By forcing a global variable (`window.search = {}`), and then stashing a
searchInstance there after load, we can let the UI object check for it
prior to instantiating a new object (and making the corresponding new
network request).