Add biblio links to catalogue.yml file
This commit adds reference links to the "Bibliography" and "Parallels"
sections of the main catalogue entry texts.

The script to do this has been included as well. It will need to be
modified somewhat to work on normal essay pages (which are not
structured as YAML data), but the basic principle should be the same.