Explicitly include 252 bis in frontmatter
The "bis" entries and other non-numerical cat numbers need to be
explicitly included in the front matter cat arrays if we want the
objects to show up in the grid. These values should be added as strings,
exactly as they appear in the catalogue.yml file (spaces, case, etc). So
"252 bis", "626a", etc.