Eliminate sub-catalogue structure and renumber
With the introductions for the “Loeschcke type VIII” section  copied
into their specific entries (7d36b25 and 4f2ddfd), there is no longer a
need for the “catalogue within a catalogue” structure, so this deletes
that from the Typology.

This commit also makes some fixed to the numbering so that it is
continuous across a couple of the sections as indicated in the
transmittal docs.