Disable SmoothState for now
I'm disabling this for the moment, as it's starting to cause too many
problems with regards to the modal object views, hash changes, etc. The
JS-enhanced functionality is regressing to be worse than the default
browser functionality. If we abandon smoothstate then some other UI
things in the JS should change (client side search, in particular). But
that can happen later.