Add sorted_concordance_entries method
This is the first stab at allowing for sorting of Concordance entries based on
inventory numbers. Some "lookup" values had to be added to the YAML data for cat
entries with non-standard values for inv numbers, and two helper methods have
been added to help with a complex sort operation.