Implement leaflet map inside of ImageViewer component
The ImageViewer component, which is a child of the Details component, is now
sucessfully rendering a leaflet map for whichever entry has been selected.

Leaflet was not rendering properly on initial load into the off-canvas DOM
elements, so the best solution seems to be to load data immediately (XHR
request, set values, etc), but to delay rendering until the "visible" state of
the parent element has been activated. This seems to work fine everywhere.