Search component with localstorage now working
This is now a standard Vue component, using the Watch method for populating its
results. The contents list (which the Lunr results are cross-referenced against
for a human-readible result) is stashed in LocalStorage when the component first
loads. Thereafter it checks localStorage first, and only loads the JSON again if
necessary.