Move Lunr setup into global Application JS from Search component
In a perfect world this wouldn't be necessary, and the Search component would
manage all of this data internally.

However, this is a hybrid application and some concessions had to be made to
performance. Parsing and loading the large amounts of JSON data required to
create a search index for this publication requires eliminating expensive and
redundant operations, even at the cost of better code architecture.