Add Catalogue entries to search index
The search feature now supports searching inside of catalogue entries in
addition to content pages. This has caused the search index to more than double
in size, so hopefully it is worth it.

The search.json builds an index that contains an accurate URL for each catalogue
entry, followed by an anchor link to its cat number. The next step is to cause
this URL value to trigger the modal catalogue detail state.