Add basic dropdown to Catalogue component
Pulling this from GeoJSON at component initialization. To save some effort down
the road, locationsList is an array of objects which contain "name", "type", and
"entries" properties. Hopefully stashing the data here when the array is
generated will save us from having to do complicated look-ups later - for
example, instead of computing which catalogue entries go with which entries, we
can have the "value" attribute of the dropdown just pass the array of relevant
entries directly, saving ourselves some code.