Use localforage for storing data
Localforage is JS library that exposes an API similar to the native LocalStorage
api, but with a number of improvments. It supports async operations, which means
less UI lag. Also, it uses IndexedDB on the back-end (at least in most browsers)
which gets around the localStorage size limits (these vary by browser, and can
also be set by the user in some browsers). The library also removes the need to
manually call JSON.parse() and JSON.stringify().

Another side-effect of this change is better code organization. Stashing data in
localForage is better than storing it in globals off of the window object (only
the Search Index object still gets this treament for now, since we need to call
methods on it). All data setup is now handled at the start of application.js, so
that things are ready by the time they are needed in other components. The
getData() method that appears in some components is now just a wrapper around
localforage.getItem(), doing any additional work in callbacks (no need to store
URLs and choreograph ajax requests across multiple places in this app).

All in all I'm pretty impressed by how easy it was to use this library; the app
is noticeably more responsive afterwards as well (and works in Chrome, Firefox,
and Safari - as well as the Ubuntu versions of the first two browsers). Will
test on IE next.