Use plurals for multiple inventory numbers
Adds a `multiple_objects: true` field and value to `catalogue.yml`
where one catalogue entry has multiple inventory numbers. When the
field is present, templates write “Inventory Numbers”, rather than
“Inventory Number”.

Collection links in these instances are then to the first object in the
series. Not ideal, but the best solution.