Change the way we handle links to the collection
Previously, on the object pages, the inventory number was linked to the
collection pages, and on the concordance, the DOR ID was listed and
linked to the collection pages. To make these consistent with one
another and with previous collection catalogues, there is now a link
icon by the inventory number, and the icon itself is the link. These
icons then don’t show up in the pdf/print edition.