Refactor JS to avoid some race conditions, improve responsiveness, fix certain bugs
This should solve the issue identified in ticket #60, and generally improves the
setup/teardown, and state management in the various JS components here.

This required significant re-organization of the whole app - a top-level page
object is now used as the namespace for other global variables and methods.
Wherever possible, async methods are used (in places like localforage) with
events emitted when data is ready so that other components don't try to access
data which doesn't exist.