Use source serif pro for greek inscriptions
This fixes an issue where the Greek characters would display on the screen (in
Mac OSX Preview or in the browser) but would fail to show up in the PDF when
printed. Requiring the font file explicitly (instead of relying on a
locally-available font name) seems to be the best way to get consistent font
rendering here.