Catalogue filter improvements
Some entries were missing values for date_numeric in the catalogue data file,
which means they were not showing up in the catalogue grid.

For now, all the objects which are truly undetermined have been given values of
-800 / 800, meaning they will *always* show up in the results. This might be
changed depending on the desired behavior. This includes the non-entry, cat 456.

Additionally, the date filter has been adjusted to be more inclusive or
"grabby". If *either* date in its date range falls within the current selection,
the object will now appear in the results.

These changes mean that the catalogue grid now correctly displays all 634
objects from the book on initial load.