Handle 'bis' entries consistently
Throughout the book, the 'bis' entries are now being treated as 146-bis (single
hyphen separating). But they need to be displayed to the user as '146 bis' (with
a space).

Frontmatter, CSS, etc. has been updated to look for the hyphenated usage. But
the catalogue grid and details grid components have been updated to use a Vue.js
filter (this is a filter in the narrower 2.0 sense, designed for text
manipulation only). The filter replaces hypens in cat numbers with a space, and
is only called where the cat_no field is directly echoed into a template.