Handle race condition with catalogue data (#80)
This commit addresses a race condition that can occur when the user visits the
catalogue component on a cold page load (no prior prep of caches, localForage,
etc).

The catalogue grid vue component how checks for window.page._catalogueStatus in
the getData() method which is called as soon as it mounts. If the
_catalogueStatus is false (i.e. has not been loaded yet) then no attempt is made
to load the data, and the component's "ready" status remains set to false as
well.

When the catalogue status is set to true, an event is broadcast globally
announcing the update. An event listener has been added to the window object (in
ui.js) which listens for this event and calls the grid component's getData()
method again to load the data once it is ready.