Fix problem with "stale" background images in catalogue grid
This fixes a side-effect of using Lazysizes and the 'attrchange' plugin in the
catalogue grid (vue component). DOM elements that were being managed by Vue in a
reactive update cycle were having their old background images "stick" until the
new one was loaded. This means that on slower connections especially, users
would start seeing images that did not properly match with the cat numbers of
updated elements in the filtered results in the grid. This was a temporary
problem but it could last long enough to be noticeable, especially on a slow
connection.

The solution was to add a global event listener for the 'lazyunveilread' event
that is used in the Lazysizes attrchange plugin. This listener calls a function
that removes the inline style attribute (added by the previous call to
Lazysizes) until the new background image is loading, so that default loading
image appears here instead.