Allow additional user-specified images to be copied into Epub (#94)
This adds an option called `additional_epub_images` to the Book extension. It
can be specified in config.yml in the activate :book block. Ex:

activate :book do |b|
  b.additional_epub_images = [
    "assets/images/fig_01.jpg",
    "assets/images/fig_02.jpg",
    "assets/images/fig_03.jpg",
    "assets/images/fig_04.jpg",
    "assets/images/cc-by.png",
    "assets/images/cover.jpg"
  ]
end

The coppy_additional_image_assets method has been added to the Epub class to
support this. Right now this is just a quick and dirty duplication of the other
copy image methods here, but since this is a one-off project there might not be
justification for re-writing these methods into a consistent single interface.