Disable toc_nav (#95)
This commit disables the generation of a "Table of Contents" page from the
build_toc_nav method. The user-defined contents page will still be added to the
epub normally, and a toc.ncx file is still being generated as well.

If this is the desired behavior I can remove this method from the codebase
instead of just commenting it out.