Polyfill the CustomEvent API for Internet Explorer (#101)
This application makes use of the Custom Event API to communicate across a
few mostly-separate parts when needed. A large amount of supporting JSON
data needs to be stored before the app can function fully. When this data
is ready for use by other components (details view, search tool, etc),
an event is broadcast announcing that the relevant data is ready.

Internet Explorer does not implement this API the same as other browsers
so this polyfill is necessary there.

The lack of IE support for this event was causing the DeepZoom component to not
get setup properly on initial page load in a catalogue entry page (the FindEntry
method would never get called until the cat property changed).

The localforage setup has also been re-written for clarity but it was not
actually the cause of the bug.